Moved torcs_server_config to utils
